web-animations/animation-effect/animation-effect-iteration-composite.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationEffect firstChild attribute test</title>
<meta name="assert" content="The iteration composite operation property of this animation effect as specified by one of the IterationCompositeOperation enumeration values">
<link rel="help" href="https://web-animations.github.io/web-animations-spec/bikeshed/Overview.html#dom-animationeffect">
<link rel="author" title="Aleksei Yu. Semenov" href="mailto:a.semenov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var effectsToTest = [newKeyframeEffect(), newMotionPathEffect()];
    effectsToTest.forEach(function(effect) {
        assert_equals(effect.iterationComposite, 'replace', 'Wrong default value of ' +
        'iterationComposite attribute of effect ' + type(effect));
    });
}, 'Test default value of iterationComposite attribute');
</script>
</body>
